# docs/cyrnel/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cyrnel API",
    "description": "Cyrnel is a universal layer that connects AI agents and LLM applications to any external service, API, or device regardless of protocol or standard. It acts as an adaptive bridge between your AI and the outside world, enabling seamless integrations through code execution, async operation handling, and built-in observability and security controls.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9371"
    }
  ],
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message returned by the API."
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code (e.g. invalid_cursor, cursor_expired)."
          }
        },
        "required": [
          "error"
        ],
        "description": "Standard error envelope returned by the HTTP error middleware."
      },
...
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "description": "Module identifier."
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Module display name."
          },
          "type": {
            "type": "string",
            "enum": [
              "adapter",
              "environment"
            ],
            "description": "Module type."
          },
          "description": {
            "type": "string",
            "description": "Human-readable description of the module."
          },
          "version": {
            "type": "string",
            "minLength": 1,
            "example": "2.0.0",
            "description": "Semantic version of the installed module manifest."
          },
          "isBuiltin": {
            "type": "boolean",
            "description": "Whether the module is bundled with the API."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the module is currently enabled."
          },
          "missing": {
            "type": "boolean",
            "description": "Whether the module is installed but has no matching factory loaded."
          },
          "hasIcon": {
            "type": "boolean",
            "description": "Whether the module has a registry-declared icon."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO-8601 timestamp of when the module was installed."
          },
          "autoUpdate": {
            "type": "boolean",
            "description": "Whether the module opted into the background auto-update sweep."
          },
          "autoUpdateConstraint": {
            "type": "string",
            "nullable": true,
            "description": "Semver range pinning the auto-update registry version, or null for latest."
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "description",
          "version",
          "isBuiltin",
          "enabled",
          "missing",
          "hasIcon",
          "createdAt",
          "autoUpdate",
          "autoUpdateConstraint"
        ],
        "description": "Module manifest record returned by the modules endpoints."
      },
      "ModuleListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            },
            "description": "Modules that match the supplied query filters."
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Opaque cursor for the next page; null when there are no more items. Echo it back as the cursor query parameter to continue paginating."
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether additional pages exist. When true, nextCursor is a valid cursor; when false, pagination has reached the end."
          }
        },
        "required": [
          "items",
          "nextCursor",
          "hasMore"
        ],
        "description": "Paginated collection envelope."
      },
      "ModuleDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Module"
          },
          {
            "type": "object",
            "properties": {
              "hash": {
                "type": "string",
                "minLength": 1,
                "description": "Content hash of the installed module archive."
              },
              "source": {
                "type": "string",
                "minLength": 0,
                "description": "Install source URL used to fetch the module archive."
              },
              "configSchema": {
                "type": "object",
                "additionalProperties": {
                  "nullable": true
                },
                "description": "JSON Schema describing the module configuration object."
              },
              "secretsSchema": {
                "type": "object",
                "additionalProperties": {
                  "nullable": true
                },
                "description": "JSON Schema describing the module secrets object."
              }
            },
            "required": [
              "hash",
              "source",
              "configSchema",
              "secretsSchema"
            ]
          }
        ],
        "description": "Full module manifest record returned by the get module endpoint."
      },
...
      "ModuleDirectInstallRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1,
            "description": "Direct URL of the .tar.zst module archive to download and install."
          },
          "autoUpdate": {
            "type": "boolean",
            "description": "Whether to auto-update this module with the archive URL it was installed from. Defaults to true."
          }
        },
        "required": [
          "url"
        ],
        "description": "Request body used to install a module from a direct URL."
      },
...
      "RateLimitErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "rate_limit_exceeded"
            ],
            "description": "Error code identifying rate-limit rejection."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message with retry instructions."
          },
          "retryAfter": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "description": "Number of seconds to wait before retrying."
          }
        },
        "required": [
          "error",
          "message",
          "retryAfter"
        ],
        "description": "Error body returned when a request is rate-limited (HTTP 429)."
      },
...
    },
...
  },
  "paths": {
...
    "/modules": {
      "get": {
        "tags": [
          "Modules"
        ],
        "summary": "List modules",
        "description": "Returns module manifests filtered by optional query, type, isBuiltin, enabled, and missing query parameters.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Free-text query used to match module names and descriptions."
            },
            "required": false,
            "description": "Free-text query used to match module names and descriptions.",
            "name": "query",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "adapter",
                "environment"
              ],
              "description": "Module type. Either 'adapter' or 'environment'."
            },
            "required": false,
            "description": "Module type. Either 'adapter' or 'environment'.",
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Filter by whether the module is bundled with the API."
            },
            "required": false,
            "description": "Filter by whether the module is bundled with the API.",
            "name": "isBuiltin",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Enabled-state filter for modules."
            },
            "required": false,
            "description": "Enabled-state filter for modules.",
            "name": "enabled",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Filter by whether the module is missing its factory."
            },
            "required": false,
            "description": "Filter by whether the module is missing its factory.",
            "name": "missing",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 2048,
              "description": "Opaque pagination token returned as nextCursor by the previous response. Pass it back unchanged to fetch the next page; null/omitted fetches the first page. Cursors encode position, not filters, so change filters only between pages when starting over."
            },
            "required": false,
            "description": "Opaque pagination token returned as nextCursor by the previous response. Pass it back unchanged to fetch the next page; null/omitted fetches the first page. Cursors encode position, not filters, so change filters only between pages when starting over.",
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "Maximum number of items per page. Clamped to a maximum of 100; defaults to 20."
            },
            "required": false,
            "description": "Maximum number of items per page. Clamped to a maximum of 100; defaults to 20.",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching modules.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleListResponse"
                }
              }
            }
          },
          "400": {
            "description": "One or more query parameters could not be parsed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "A bearer token was required but missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. The request was throttled. Check Retry-After header.",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Maximum requests allowed in the current window."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in the current window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp when the window resets."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait before retrying."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The module list could not be loaded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Modules"
        ],
        "summary": "Install a module from a direct URL",
        "description": "Downloads a .tar.zst archive from the supplied direct URL, validates the module.json manifest, and registers the module. Newly installed modules start disabled by default. No registry source is stored.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleDirectInstallRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The module manifest record for the newly installed module.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetails"
                }
              }
            }
          },
          "400": {
            "description": "The request body was invalid or the archive failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "A bearer token was required but missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "A module already exists with the requested id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The downloaded archive exceeded the configured size limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. The request was throttled. Check Retry-After header.",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Maximum requests allowed in the current window."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in the current window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp when the window resets."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait before retrying."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The module could not be installed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The archive file could not be downloaded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}